+++ Source/Aul2AudioViewPlugin.pas
@@ -27,7 +27,7 @@ uses
 
 var
   GViewTypeSelect  : TFILTER_ITEM_SELECT;
-  GViewTypeList    : array[0..12] of TFILTER_ITEM_SELECT_ITEM;
+  GViewTypeList    : array[0..13] of TFILTER_ITEM_SELECT_ITEM;
   GViewStyleSelect : TFILTER_ITEM_SELECT;
   GViewStyleList   : array[0..2] of TFILTER_ITEM_SELECT_ITEM;
   GViewDensityTrack: TFILTER_ITEM_TRACK;
@@ -188,6 +188,7 @@ begin
     AddSelectList(GViewTypeList, 'Radial Waveform (3D)', VIEW_TYPE_RADIAL_WAVEFORM_3D);
     AddSelectList(GViewTypeList, 'Spectrum Landscape (3D)', VIEW_TYPE_SPECTRUM_LANDSCAPE_3D);
     AddSelectList(GViewTypeList, 'Waveform Tunnel (3D)', VIEW_TYPE_WAVEFORM_TUNNEL_3D);
+    AddSelectList(GViewTypeList, 'Spectrum Waterfall (3D)', VIEW_TYPE_SPECTRUM_WATERFALL_3D);
     AddSelect(GViewTypeSelect, 'Type', VIEW_TYPE_EQUALIZER_BARS, @GViewTypeList[0]);
 
     ClearSelectList;